+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@0.5.0...@amplitude/analytics-node@0.5.1) (2022-09-16)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
 # [0.5.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@0.4.1...@amplitude/analytics-node@0.5.0) (2022-09-08)