+++ packages/analytics-types/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-types@0.9.0...@amplitude/analytics-types@0.10.0) (2022-09-16)
+
+
+### Features
+
+* new marketing analytics plugin ([#213](https://github.com/amplitude/Amplitude-TypeScript/issues/213)) ([02ff174](https://github.com/amplitude/Amplitude-TypeScript/commit/02ff174e3361173dbf15ed3acf72e950810e174f))
+
+
+
+
+
 # [0.9.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-types@0.8.1...@amplitude/analytics-types@0.9.0) (2022-09-08)